classfmt: fix "Implicit narrowing conversion" warnings
ClassFileStruct.u4At(int) is supposed to return an unsigned integer. The
return value is currently always used as primitive integer (signed) to
compute readOffset inside the class file (which will then passed as
argument to the method again). So any negative value would trigger an
ArrayIndexOutOfBoundsException when used as index into the byte array.

The intermediate usage of primitive long did not help and only triggered
warnings at the call sites.

https://github.com/eclipse-jdt/eclipse.jdt.core/security/code-scanning/18
ff

+++ org.eclipse.jdt.core.compiler.batch/src/org/eclipse/jdt/internal/compiler/classfmt/MethodInfo.java
@@ -482,7 +482,7 @@ private synchronized void readCodeAttribute() {
 }
 private void decodeCodeAttribute(int offset) {
 	int readOffset = offset + 10;
-	int codeLength = (int) u4At(readOffset);
+	int codeLength = u4At(readOffset);
 	readOffset += (4 + codeLength);
 	int exceptionTableLength = u2At(readOffset);
 	readOffset += 2;
